refactor: WorkScreen 컴포넌트를 ExhibitionWorkPage로 통합 및 파일 정리
